Add files via upload
The Auto-Spare Part Management is developed for sale/purchase organization. This system manages customer, product, sales order, and dealer and purchase order. It provides a platform to the trader to automate their record keeping. The purpose of this project is to develop an application program to reduce the manual work for managing inventory, product, customer, dealer and orders. This project is built to help admin in managing his organization. Admin can store customer details in the database. Next time when customer comes in then there is no need to enter his details again. It can selected from the list displaying all customer